Add tests for escape characters in text
1. Test that the following inputs are escaped and formatted properly.
	\c		No \tab
	\`		\`Not Raw`
	\\		\\`Raw`
	\#		\#Not Header
	\-		\--Not Strikethrough--
	\*		\***No Format***
	\[		\[Not Cite]
	\]		[Not Cite\]
	\(		[Direct Cite]\(Not Sourced)
	\)		[Direct Cite](Not Sourced\)
	\_		\_Not underlined_
2. Currently, str in parser does not return raw input on eof. This
prevented a line beginning with a '[' from defaulting to a paragraph if
the closing ']' is not reached (or if it is escaped). We now maintain
two separate buffers to return the original unskipped runes as a string.
3. Currently, as a consequence of having a "](" token, inputs that don't
close with a ')' don't parse a direct citation for the text inside the
"["..."]". We now handle this case by checking that there is an
unhandled "["..."](" in the cit slice, and return an ast.Format object
accordingly.