gen/html: Refactor API to be similar to os/exec's
1. Refactored API for html generation to be similar to that of os/exec
	a) Previously, the Genner struct implemented io.Reader and
io.WriterTo, so a constructed &html.Genner{} could be passed into
utility functions like io.Copy. The utility of this API was limited
because Stderr and Stdout couldn’t be sent to the same io.Writer, unless
they were both sent to Stdout in different ways.
	b) The new API is more flexible since one can set the Writer’s via
the struct, connect a pipe to one of them, slurp all the output at once,
and start and block for completion at separate times.
2. Added docs and examples for html package
	a) Almost every top-level function in the html package has an
example.