parser: rename parse->parser, add MustParse & docs
package parse has been renamed to parser, because parser.Parse makes
more sense from a "subject-verb" perspective than parse.Parse.

MustParse has been added as a top-level function to make easier the global
initialization of an *ast.File as well as situations where the input
source is known to be valid.

Package documentation is improved to specify the grammar and escape
characters in source files.