parser: ListItem separation, examples, more tests
Fixed issue with list items where a new list was not being
created when there was an empty line. Now, if a successive
line is empty or prefixed with "-", we concatenate to our
accumlator-line an eof, the successive line, and a newline
character. Otherwise, we simply concatenate a single space
character and the successive line.

    acc = line()
    succ = line()
    for !empty(succ) and !prefixed(succ, "-"):
        acc += " " + succ
    acc += eof + succ + "\n"

Parser now has tests to verify that
    1) ListItem seaparation works properly.
    2) Paragraph combination works properly.
    3) Unicode code points are handled correctly w.r.t. formatting.
        a) Language tests include Telugu, Arabic, Hebrew,
           and Chinese (Simplified).
        b) Also test mathematical symbols and emoji.

Parser documentation includes an example for MustParse
that simply parses a bulleted list (LOTR references).