mexdown: Add command line utility in mexdown.go
This change adds a command line utility in mexdown.go to read an input file
an run a generator (for a subcommand) that writes to an output
file.

Right now, since we only have an HTML library generator, the CLI only
has an HTML generator registered.

Updated go.mod to include cobra and pflag.