parser: Preserve all newlines for ast.Paragraph
Currently the parser stops parsing a paragraph at the end of a line or
EOF. When merging consecutive paragraphs, they are artificially joined
with a newline. However, users that want multiple newlines between
paragraphs have to resort to nbsp; or some other hack to get there.

This change modifies the parser to parse newlines into the paragraph.
Combining consecutive paragraphs is simply a concatenation of the two
bodies of text.